+++ src/ui/modules/auth/pages/AuthPage/RegisterForm.jsx
