+++ src/ui/modules/customer/components/Carousel/Carousel.css
