+++ src/ui/modules/customer/components/Divider/Divider.jsx
