+++ src/ui/modules/customer/pages/Register/Register.jsx
