+++ src/ui/modules/provider/components/DashboardLayout/DashboardLayout.css
